refactor: 라우팅 관련 로직 헬퍼 분리 #26
